+++ CHANGELOG.md
@@ -60,8 +60,9 @@ Highlights are marked with a pancake 🥞
 - store: Use `AnyOperation` on `LogStore` [#1391](https://github.com/p2panda/p2panda/pull/1391)
 - sync: Update examples to use new `api` module [#1413](https://github.com/p2panda/p2panda/pull/1413)
 
-### Fixes
+### Fixed
 
+- core: Make `cbor_decode` lenient [#1429](https://github.com/p2panda/p2panda/pull/1429)
 - node: Allow event processing to handle out-of-order buffering by separating i/o streams and preserve input ordering [#1271](https://github.com/p2panda/)
 - node: Unblock task tracker by introducing "pass-through" events coming from orderer [#1267](https://github.com/p2panda/p2panda/pull/1267)
 - spaces: Deterministic deserialization of `SpacesArgs` [#1264](https://github.com/p2panda/p2panda/pull/1264)